parser_service/requirements.txt
grpcio==1.68.1
grpcio-tools==1.68.1
psycopg2-binary== 2.9.10
pika==1.3.2
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.27.1
webdriver-manager==4.0.2
sqlalchemy==2.0.36
alembic==1.14.0